+++ mlflow/pyfunc/utils.py
@@ -1,8 +1,7 @@
-import json
 import os
 import sys
-from mlflow.utils.rest_utils import NumpyEncoder
-from mlflow.utils import get_jsonable_obj
+
+
 
 def _add_code_to_system_path(code_path):
     sys.path = [code_path] + _get_code_dirs(code_path) + sys.path
