+++ mlflow/utils/__init__.py
@@ -10,26 +10,6 @@ PYTHON_VERSION = "{major}.{minor}.{micro}".format(major=version_info.major,
                                                   minor=version_info.minor,
                                                   micro=version_info.micro)
 
-
-def get_jsonable_obj(data, pandas_orient="records"):
-    """Attempt to make the data json-able via standard library.
-    Look for some commonly used types that are not jsonable and convert them into json-able ones.
-    Unknown data types are returned as is.
-
-    :param data: data to be converted, works with pandas and numpy, rest will be returned as is.
-    :param pandas_orient: If `data` is a Pandas DataFrame, it will be converted to a JSON
-                          dictionary using this Pandas serialization orientation.
-    """
-    if isinstance(data, np.ndarray):
-        return data.tolist()
-    if isinstance(data, pd.DataFrame):
-        return data.to_dict(orient=pandas_orient)
-    if isinstance(data, pd.Series):
-        return pd.DataFrame(data).to_dict(orient=pandas_orient)
-    else:  # by default just return whatever this is and hope for the best
-        return data
-
-
 def get_major_minor_py_version(py_version):
     return ".".join(py_version.split(".")[:2])
